+++ Convertor/ConvertWrapper.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Yireo\Webp2\Convertor;
 
+use Exception;
 use Psr\Log\LoggerInterface;
 use WebPConvert\Convert\Exceptions\ConversionFailed\InvalidInput\InvalidImageTypeException;
 use WebPConvert\Convert\Exceptions\ConversionFailedException;
@@ -52,8 +53,8 @@ class ConvertWrapper
             try {
                 WebPConvert::convert($sourceImageFilename, $destinationImageFilename, $options);
                 break;
-            } catch (\Exception $e) {
-                $this->logger->debug($e->getMessage() . ' - ' . $e->description, $e->getTrace());
+            } catch (Exception $e) {
+                $this->logger->debug($e->getMessage(), $e->getTrace());
                 continue;
             }
         }